pyproject.toml
# ******************************************************************************
#
# metietur, components for measurement practice
#
# Copyright 2021-2024 Jeremy A Gray <gray@flyquackswim.com>.
#
# All rights reserved.
#
# SPDX-License-Identifier: Proprietary
#
# ******************************************************************************

[build-system]

requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.bandit.assert_used]

exclude_dirs = [
  ".git",
  ".pytest_cache",
  ".tox",
  ".venv",
  "dist",
  "docs",
  "htmlcov",
]

skips = [
  "**/test_*.py",
  "tests/",
]

[tool.black]

line-length = 88
include = '\.pyi?$'
exclude = '''

(
  /(
    \.git
    | \.pytest_cache
    | __pycache__
    | htmlcov
  )/
)
'''

[tool.doc8]

ignore-path = [
  ".tox",
  ".venv",
]

[tool.isort]

ensure_newline_before_comments = true
force_single_line = true
include_trailing_comma = true
line_length = 88
multi_line_output = 3
use_parentheses = true

[tool.pccc]

header_length = 50
body_length = 72
wrap = true
force_wrap = true
spell_check = false
repair = false
ignore_generated_commits = false

types = [
  "depends",
  "doc",
  "feat",
  "fix",
  "perf",
  "refactor",
  "release",
  "style",
  "test",
]

scopes = [
  "js",
  "py",
  "tooling",
]

[tool.poetry]

authors = [
  "Jeremy A Gray <gray@flyquackswim.com>",
]
description = "components for measurement practice"
name = "metietur"
version = "0.0.1"
license = "GPL-3.0-or-later"
maintainers = [
  "Jeremy A Gray <gray@flyquackswim.com>",
]
readme = "README.md"
homepage = "https://github.com/jeremyagray/metietur"
repository = "https://github.com/jeremyagray/metietur"

classifiers = [
  "Development Status :: 2 - Pre-Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]

packages = [
  { include = "metietur" },
]

[tool.poetry.dependencies]

python = ">=3.10.1,<4.0"
svg-py = ">=1"

[tool.poetry.group.dev.dependencies]

bandit = ">=1"
black = ">=24"
doc8 = ">=1"
flake8 = ">=7"
hypothesis = ">=6"
isort = ">=5"
pccc = ">=0"
pre-commit = ">=3"
pydocstyle = ">=6"
pytest = ">=8"
pytest-cov = ">=5"
sphinx = ">=8"
tox = ">=4"
yesqa = ">=1"

[tool.poetry.scripts]

clock = "metietur:main"

[tool.poetry.urls]

"Issues" = "https://github.com/jeremyagray/metietur/issues"
"Documentation" = "https://metietur.readthedocs.io/"
"Repository" = "https://github.com/jeremyagray/metietur"

[tool.pytest.ini_options]

pythonpath = [
  ".",
]
